Cleanup IODE:     - suppress unreferenced local vars     - declare missing function prototypes     - transform old C declarations in C++ std     - cast strlen() to int     - move static functions above calling fns     - add all fn declarations (via mkh) on top of the code     - correct void functions returning a value     - delete unreferenced labels

+++ api/buf.c
@@ -113,5 +113,5 @@ char *BUF_memcpy(char *ptr, int lg)
 
 char *BUF_strcpy(char *ptr)
 {
-    return(BUF_memcpy(ptr, strlen(ptr) + 1));
+    return(BUF_memcpy(ptr, (int)strlen(ptr) + 1));
 }
